cp: serve files from cache
Don't stream the media files but instead download them into a managed cache. Reason being that e.g. WhatApp that is opening the media files tries to open the ParcelFileDescriptor as "whole". I.e. it seems that when it's not possible the read() the whole PFD in one go, then WhatsApp fails. It anyway opened the media file twice. Once probably to check the file type or other stuff, and then a second time for actual opening.